basic plotting
It's a bit rough around the edges, but this gives us a basic way to
record the energy state at various times during each run.

Observation of the data: The interesting thing here is that for
simulated annealing, we get a chart where the same number is reached
multiple times toward the end of the run, This is because the state is
mutated toward a higher energy level (sometimes) even toward the end of
the run.