make program multithreaded
This was almost too easy with the use of the excellent parallel-io
library (and the fact that we have a classic "embarrassingly parallel"
problem of repeating the same operation, `genLayoutScramble', N times).
The only "gotcha" was that we had to use the line

    Control.Concurrent.setNumCapabilities threads

because that's how the GHC RTS (runtime system) deals with concurrency
(multiple threads).